Pull common code out of *_merge tools, fixes bug in BIN_merge
