+++ testing/test_tools.py
@@ -4,6 +4,7 @@
 # Run tests and show captured stdout:
 # python -m pytest -rP
 
+import BIN_merge
 import map_maker
 import show_types
 import tlog_info
@@ -16,9 +17,22 @@ import tlog_scan
 
 class TestTools:
 
+    def test_BIN_merge(self):
+        tool = BIN_merge.DataflashLogReader('testing/small.BIN', ['VIBE'], 10000, 10000, False)
+        tool.read()
+        tool.write_merged_csv_file()
+
     def test_map_maker(self):
         map_maker.build_map_from_tlog('testing/small.tlog', 'testing/small.html', False, [None, None], 18,
                                       ['GLOBAL_POSITION_INT'], 10.0)
+
+    def test_tlog_types(self):
+        tool = show_types.TypeFinder('testing/small.tlog')
+        tool.read()
+
+    def test_BIN_types(self):
+        tool = show_types.TypeFinder('testing/small.BIN')
+        tool.read()
 
     def test_tlog_info(self):
         tool = tlog_info.TelemetryLogInfo('testing/small.tlog')
@@ -27,6 +41,7 @@ class TestTools:
     def test_tlog_merge(self):
         tool = tlog_merge.TelemetryLogReader('testing/small.tlog', ['GLOBAL_POSITION_INT'], 10000, 10000, False, True)
         tool.read_tlog()
+        tool.write_merged_csv_file()
 
     def test_tlog_param(self):
         tool = tlog_param.TelemetryLogParam()
@@ -35,8 +50,4 @@ class TestTools:
 
     def test_tlog_scan(self):
         tool = tlog_scan.Scanner('testing/small.tlog', ['GLOBAL_POSITION_INT'])
-        tool.read()
-
-    def test_tlog_types(self):
-        tool = show_types.TypeFinder('testing/small.tlog')
         tool.read()
