+++ tlog_merge.py
@@ -6,46 +6,39 @@ operation does a forward-fill (data is copied from the previous row), so the res
 substantially larger than the sum of the per-type csv files.
 """
 
-import os
 from argparse import ArgumentParser
 
-import pandas as pd
 from pymavlink import mavutil
 
 import table_types
 import util
+from log_merger import LogMerger
 
 
-class TelemetryLogReader:
+class TelemetryLogReader(LogMerger):
     def __init__(self,
-                 tlog_filename: str,
+                 infile: str,
                  msg_types: list[str],
                  max_msgs: int,
                  max_rows: int,
                  verbose: bool,
-                 all: bool):
-        self.tlog_filename = tlog_filename
-        self.prefix = tlog_filename.split('.')[0]
-        self.msg_types = msg_types
-        self.max_msgs = max_msgs
-        self.max_rows = max_rows
-        self.verbose = verbose
-        self.all = all
-        self.tables = None
+                 all_types: bool):
+        super().__init__(infile, msg_types, max_msgs, max_rows, verbose)
+        self.all_types = all_types
 
     def read_tlog(self):
         self.tables = {}
         for msg_type in self.msg_types:
             self.tables[msg_type] = table_types.Table.create_table(msg_type, self.verbose)
 
-        print(f'Reading {self.tlog_filename}')
-        mlog = mavutil.mavlink_connection(self.tlog_filename, robust_parsing=True, dialect='ardupilotmega')
+        print(f'Reading {self.infile}')
+        mlog = mavutil.mavlink_connection(self.infile, robust_parsing=True, dialect='ardupilotmega')
 
         print('Parsing messages')
         msg_count = 0
         while (msg := mlog.recv_match(blocking=False, type=self.msg_types)) is not None:
             # Only consider messages from ArduSub
-            if not self.all and (msg.get_srcSystem() != 1 or msg.get_srcComponent() != 1):
+            if not self.all_types and (msg.get_srcSystem() != 1 or msg.get_srcComponent() != 1):
                 continue
 
             msg_type = msg.get_type()
@@ -68,50 +61,6 @@ class TelemetryLogReader:
                 print(f'{msg_count} messages')
 
         print(f'{msg_count} messages')
-
-    def outfile(self, suffix: str = '', ext: str = '.csv'):
-        dirname, basename = os.path.split(self.tlog_filename)
-        root, _ = os.path.splitext(basename)
-        return os.path.join(dirname, root + suffix + ext)
-
-    def write_msg_csv_files(self):
-        print('Writing csv files')
-        for msg_type in self.msg_types:
-            df = self.tables[msg_type].get_dataframe(self.verbose)
-            if len(df):
-                filename = self.outfile(suffix=f'_{msg_type}')
-                print(f'Writing {len(df)} rows to {filename}')
-                df.to_csv(filename)
-
-    def write_merged_csv_file(self):
-        merged_df = None
-        print(f'Merging dataframes')
-        for msg_type in self.msg_types:
-            df = self.tables[msg_type].get_dataframe(self.verbose)
-            if df.empty:
-                if self.verbose:
-                    print(f'{msg_type} empty, skipping')
-            else:
-                if merged_df is None:
-                    if self.verbose:
-                        print(f'Starting with {len(df)} {msg_type} rows')
-                    merged_df = df
-                else:
-                    if self.verbose:
-                        print(f'Merging {len(df)} {msg_type} rows')
-                    merged_df = pd.merge_ordered(merged_df, df, on='timestamp', fill_method='ffill')
-                    if self.verbose:
-                        print(f'Merged dataframe has {len(merged_df)} rows')
-                    if len(merged_df) > self.max_rows:
-                        print('Merged dataframe is too big, stopping')
-                        break
-
-        if merged_df is None:
-            print(f'Nothing to write')
-        else:
-            filename = self.outfile()
-            print(f'Writing {len(merged_df)} rows to {filename}')
-            merged_df.to_csv(filename)
 
 
 def main():
